feat: enable short jumps with quick jump btn press
pygame events on keydown and keyup record and calculate difference
between time when jump was pressed and released.
It applies opposite velocity towards the earth, if the difference is
less than, interval suitable for short jumps.
It is somewhat hacky, but will do for now. Will see how the data and
pattern for this behavior emerges and compress it further.